Create JsonRpcSigningServer
